requirements.txt
shiny
pandas
pytesseract
pdfplumber
pillow
openai
openpyxl